Location Maker
I moved the Location Maker to its own project, so to run it, set it as
the startup project. It'll boot up the android emulator too since it
references things from the app. I temporarily changed/commented out the
storage directory variable (storageDir) in the LocationManager class to
save on the desktop, so it will probably only work with emulators right
now. You'll have to edit storageDir to reflect the desktop directory of
your computer. The LocationMaker itself works by clicking on the image
and will remember Locations between sessions. The red Xs are just so we
know it's working. It still has to have Locations spearated into
separate floors/buildings. The upper left button and dropdown don't do
anything right now.